cost-savings/requirements.txt
fastapi>=0.122.0
uvicorn>=0.38.0
streamlit
requests>=2.32.5
pydantic>=2.12.4
python-dotenv>=1.2.1

# AI & Logic
langchain>=1.1.0
langchain-google-genai>=3.2.0
langchain-core>=1.1.0